requirements.txt
asn1crypto==0.24.0
blinker==1.4
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.2
django-bootstrap4==0.0.7
django-heroku==0.3.1
google-api-python-client==1.7.6
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.0
idna==2.8
oauthlib==2.1.0
psycopg2==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
